+++ atlasdb-client/src/main/java/com/palantir/atlasdb/AtlasDbConstants.java
@@ -78,8 +78,11 @@ public final class AtlasDbConstants {
     public static final int DEFAULT_TABLE_COMPRESSION_BLOCK_SIZE_KB = 8;
     public static final int DEFAULT_TABLE_WITH_RANGESCANS_COMPRESSION_BLOCK_SIZE_KB = 64;
 
+    public static final long STARTING_TS = 1L;
     public static final long TRANSACTION_TS = 0L;
     public static final long MAX_TS = Long.MAX_VALUE;
+
+    public static final byte[] DEFAULT_METADATA_COORDINATION_KEY = PtBytes.toBytes("m");
 
     public static final long DEFAULT_TRANSACTION_LOCK_ACQUIRE_TIMEOUT_MS = 60_000;
     public static final int THRESHOLD_FOR_LOGGING_LARGE_NUMBER_OF_TRANSACTION_LOOKUPS = 10_000_000;