+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/internalschema/TransactionSchemaManager.java
@@ -16,14 +16,24 @@
 
 package com.palantir.atlasdb.internalschema;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.google.common.collect.Iterables;
+import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.coordination.CoordinationService;
 import com.palantir.atlasdb.coordination.ValueAndBound;
 import com.palantir.atlasdb.keyvalue.impl.CheckAndSetResult;
+import com.palantir.logsafe.SafeArg;
 import com.palantir.logsafe.exceptions.SafeIllegalStateException;
 
 public class TransactionSchemaManager {
+    private static final Logger log = LoggerFactory.getLogger(TransactionSchemaManager.class);
+
     private final CoordinationService<InternalSchemaMetadata> coordinationService;
 
     public TransactionSchemaManager(CoordinationService<InternalSchemaMetadata> coordinationService) {
@@ -40,38 +50,70 @@ public class TransactionSchemaManager {
      * otherwise, achieving a consensus may take a long time.
      */
     public int getTransactionsSchemaVersion(long timestamp) {
+        if (timestamp < AtlasDbConstants.STARTING_TS) {
+            throw new SafeIllegalStateException("Query attempted for timestamp {} which was never given out by the"
+                    + " timestamp service, as timestamps start at {}",
+                    SafeArg.of("queriedTimestamp", timestamp),
+                    SafeArg.of("startOfTime", AtlasDbConstants.STARTING_TS));
+        }
         Optional<Integer> possibleVersion =
                 extractTimestampVersion(coordinationService.getValueForTimestamp(timestamp), timestamp);
         while (!possibleVersion.isPresent()) {
             CheckAndSetResult<ValueAndBound<InternalSchemaMetadata>> casResult = tryPerpetuateExistingState();
             possibleVersion = extractTimestampVersion(casResult.existingValues()
-                    .stream()
-                    .filter(valueAndBound -> valueAndBound.bound() >= timestamp)
-                    .findAny(),
+                            .stream()
+                            .filter(valueAndBound -> valueAndBound.bound() >= timestamp)
+                            .findAny(),
                     timestamp);
         }
         return possibleVersion.get();
     }
 
     /**
-     * Installs a new transactions table schema version, by submitting a relevant transform.
+     * Attempts to install a new transactions table schema version, by submitting a relevant transform.
+     *
+     * The execution of this method does not guarantee that the provided version will eventually be installed.
+     * This method returns true if and only if in the map agreed by the coordination service evaluated at the validity
+     * bound, the transactions schema version is equal to newVersion.
      */
-    public void tryInstallNewTransactionsSchemaVersion(int newVersion) {
-        coordinationService.tryTransformCurrentValue(valueAndBound -> {
-            if (!valueAndBound.value().isPresent()) {
-                throw new SafeIllegalStateException("Persisted value is empty, which is unexpected.");
-            }
+    public boolean tryInstallNewTransactionsSchemaVersion(int newVersion) {
+        List<Integer> presentVersionPeakValidity = coordinationService.tryTransformCurrentValue(
+                valueAndBound -> installNewVersionInMapOrDefault(newVersion, valueAndBound))
+                .existingValues()
+                .stream()
+                .map(valueAndBound -> valueAndBound.value()
+                        .orElseThrow(() -> new SafeIllegalStateException("Unexpectedly found no value in store"))
+                        .timestampToTransactionsTableSchemaVersion()
+                        .getValueForTimestamp(valueAndBound.bound()))
+                .collect(Collectors.toList());
+        return Iterables.getOnlyElement(presentVersionPeakValidity) == newVersion;
+    }
 
-            InternalSchemaMetadata internalSchemaMetadata = valueAndBound.value().get();
-            return InternalSchemaMetadata.builder()
-                    .from(internalSchemaMetadata)
-                    .timestampToTransactionsTableSchemaVersion(
-                            installNewVersionInMap(
-                                    internalSchemaMetadata.timestampToTransactionsTableSchemaVersion(),
-                                    valueAndBound.bound() + 1,
-                                    newVersion))
-                    .build();
-        });
+    private InternalSchemaMetadata installNewVersionInMapOrDefault(int newVersion,
+            ValueAndBound<InternalSchemaMetadata> valueAndBound) {
+        if (!valueAndBound.value().isPresent()) {
+            log.warn("Attempting to install a new transactions schema version {}, but no past data was found,"
+                            + " so we attempt to install default instead. This should normally only happen once per"
+                            + " server, and only on or around first startup since upgrading to a version of AtlasDB"
+                            + " that is aware of the transactions table. If this message persists, please contact"
+                            + " support.",
+                    SafeArg.of("newVersion", newVersion));
+            return InternalSchemaMetadata.defaultValue();
+        }
+
+        log.info("Attempting to install a new transactions schema version {}, on top of schema metadata"
+                + " that is valid up till timestamp {}.",
+                SafeArg.of("newVersion", newVersion),
+                SafeArg.of("oldDataValidity", valueAndBound.bound()));
+        InternalSchemaMetadata internalSchemaMetadata = valueAndBound.value().get();
+        return InternalSchemaMetadata.builder()
+                .from(internalSchemaMetadata)
+                .timestampToTransactionsTableSchemaVersion(
+                        installNewVersionInMap(
+                                internalSchemaMetadata.timestampToTransactionsTableSchemaVersion(),
+                                valueAndBound.bound() + 1,
+                                newVersion))
+                .build();
     }
 
     private TimestampPartitioningMap<Integer> installNewVersionInMap(
@@ -81,8 +123,7 @@ public class TransactionSchemaManager {
 
     private CheckAndSetResult<ValueAndBound<InternalSchemaMetadata>> tryPerpetuateExistingState() {
         return coordinationService.tryTransformCurrentValue(valueAndBound ->
-                valueAndBound.value().orElseThrow(
-                        () -> new SafeIllegalStateException("Cannot perpetuate an existing state that didn't exist!")));
+                valueAndBound.value().orElseGet(InternalSchemaMetadata::defaultValue));
     }
 
     private static Optional<Integer> extractTimestampVersion(