+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/internalschema/persistence/CoordinationServices.java
@@ -16,9 +16,17 @@
 
 package com.palantir.atlasdb.internalschema.persistence;
 
+import java.util.function.LongSupplier;
+
+import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.coordination.CoordinationService;
+import com.palantir.atlasdb.coordination.CoordinationServiceImpl;
 import com.palantir.atlasdb.coordination.TransformingCoordinationService;
+import com.palantir.atlasdb.coordination.keyvalue.KeyValueServiceCoordinationStore;
 import com.palantir.atlasdb.internalschema.InternalSchemaMetadata;
+import com.palantir.atlasdb.keyvalue.api.KeyValueService;
+import com.palantir.remoting3.ext.jackson.ObjectMappers;
+import com.palantir.timestamp.TimestampService;
 
 public final class CoordinationServices {
     private CoordinationServices() {
@@ -31,5 +39,27 @@ public final class CoordinationServices {
                 rawCoordinationService,
                 InternalSchemaMetadataPayloadCodec::decode,
                 InternalSchemaMetadataPayloadCodec::encode);
+    }
+
+    public static CoordinationService<InternalSchemaMetadata> createDefault(
+            KeyValueService keyValueService,
+            TimestampService timestampService,
+            boolean initializeAsync) {
+        return createDefault(keyValueService, timestampService::getFreshTimestamp, initializeAsync);
+    }
+
+    public static CoordinationService<InternalSchemaMetadata> createDefault(
+                KeyValueService keyValueService,
+                LongSupplier timestampSupplier,
+                boolean initializeAsync) {
+        CoordinationService<VersionedInternalSchemaMetadata> versionedService = new CoordinationServiceImpl<>(
+                KeyValueServiceCoordinationStore.create(
+                        ObjectMappers.newServerObjectMapper(),
+                        keyValueService,
+                        AtlasDbConstants.DEFAULT_METADATA_COORDINATION_KEY,
+                        timestampSupplier,
+                        VersionedInternalSchemaMetadata.class,
+                        initializeAsync));
+        return wrapHidingVersionSerialization(versionedService);
     }
 }